basic_langgraph_agent/pyproject.toml
[project]
name = "basic"
version = "0.1.0"
description = "Basic LangGraph agent for KAgent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "kagent-langgraph",
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "langchain-google-genai>=1.0.0",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "uvicorn>=0.20.0",
  "fastapi>=0.100.0",
  "httpx>=0.24.0",
  "anthropic>=0.25.0",
  "openai>=1.0.0",
  "starlette>=0.20.0",
  "sse-starlette>=1.0.0",
  "a2a-sdk"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# Add this section to tell setuptools what to include
[tool.setuptools.packages.find]
include = ["agent*"]
exclude = ["packages*"]

[project.scripts]
agent_currency = "agent.cli:main"